654: BreakSpan

+++ Typography.TextBreak/Typography.TextBreak/CustomDict.cs
@@ -211,7 +211,7 @@ namespace Typography.TextBreak
     public struct BreakSpan
     {
         public int startAt;
-        public int len;
+        public ushort len;
     }
